+++ src/components/button/index.ts
@@ -1,0 +1,3 @@
+import Button from '@/components/button/Button';
+import LoadingSpinnerButton from '@/components/button/LoadingSpinnerButton';
+export { Button, LoadingSpinnerButton };
