+++ src/components/input/index.ts
@@ -1,0 +1,4 @@
+import Input from '@/components/input/Input';
+import InputField from '@/components/input/InputField';
+
+export { Input, InputField };
