+++ src/pages/editprofile/components/CurrentPasswordInput.tsx
@@ -1,4 +1,4 @@
-import InputField from '@/components/InputField';
+import InputField from '@/components/input/InputField';
 import { useState } from 'react';
 
 interface CurrentPasswordInputPros {
