+++ src/pages/signup/components/PasswordConfirmInput.tsx
@@ -1,4 +1,4 @@
-import InputField from '@/components/InputField';
+import InputField from '@/components/input/InputField';
 import { useState } from 'react';
 
 interface PasswordConfirmInputProps {
